requirements.txt
torch==2.6.0
torchvision==0.21.0


numpy==1.26.4
plyfile
huggingface_hub
safetensors
openai>=1.12.0
google-genai>=1.0.0
httpx>=0.24.0
opencv-python>=4.8.0
Pillow>=10.0.0 
groundingdino_py
supervision==0.6.0
tqdm
ultralytics
pandas
matplotlib
scipy
scikit-learn
transformers>=4.57.0,<5
accelerate
Flask
clip
datasets
moondream
pytest
einops
rembg[gpu]
timm
